requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
transformers==4.51.3
torch>=2.2.0
openai-whisper==20240930; python_version < "3.13"
python-multipart==0.0.20
numpy>=1.26.0
librosa==0.10.2.post1
soundfile==0.12.1
